# Optimize issorted for rev=true