@LilithHafner 2024-04-23
Branching on ordering as early as possible should be beneficial for sorting runtime (though not necessarily compile time). When I benchmarked this, however, I found a surprisingly small to negligible difference.

I'd be open to merging a PR that stops telling the compiler to specialize on whether an order is reversed, but I'd want to see some motivation for it in benchmarks (e.g. decreased compile times).

Such a PR would also need to adapt existing special cases for ForwardOrdering and ReverseOrdering which are substantially faster than generic orderings in some common cases.

---

I would say the underlying issue here is the failure of constant propagation combined with not explicitly forcing a runtime branch. For example, `sort!($rs, rev=true)` happens not to have this allocation or runtime hit.