@nsajko 2024-04-23
My logic is that [`Base.Order.ord`](https://docs.julialang.org/en/v1/base/sort/#Base.Order.ord) is itself a public function, so it'd, ideally, be type-stable.